Fixed gTOwerFemaleTrainerGfxIds dtype

+++ src/battle_tower.c
@@ -176,7 +176,7 @@ const u16 gTowerMaleTrainerGfxIds[30] =
     OBJ_EVENT_GFX_HIKER
 };
 
-const u16 gTowerFemaleTrainerGfxIds[20] =
+const u8 gTowerFemaleTrainerGfxIds[20] =
 {
     OBJ_EVENT_GFX_WOMAN_2,
     OBJ_EVENT_GFX_TUBER_F,
